Renamed Model styles file
